+++ src/main/scala-2.11/markets/clearing/Fill.scala
@@ -3,7 +3,7 @@ package markets.clearing
 import java.util.UUID
 
 import markets.Message
-import markets.clearing.engines.matches.Matching
+import markets.clearing.engines.Matching
 
 
 case class Fill(matchedOrders: Matching, timestamp: Long, uuid: UUID) extends Message
